requirements.txt
python-lirc==1.2.1
paho-mqtt==1.2
